Add container-sync functionality
This commit adds containter sync functionality by allowing
the cookbook operator to override the necessary parameters
to enable the feature.  It is backwards compatible, so unless
the attributes are overriden (and they have to be, only the
site operator knows the fqdn names of the other clusters) things
will behave as before.

It is presumed that different clusters are most likely controlled
by separate and distinct chef servers since they are more then
likely in separate geographical zones, so search functionality
here would be pointless.

Change-Id: I91b25946a2ce290e63492db5d5c3fae21ef05ce5